+++ Assets/akanevrc/AnimatorControllerOverwriter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.1.0
+
+New feature and bug fix
+
+- Select modes when same name layers exist between base and overwriting
+- Fix bugs about layer name prefixes
+
 ## 1.0.0
 
 Initial release